+++ server-toss/src/modules/chance/ad-view.entity.ts
